Move RasterFileDataSource into testutils.io
This class is only used by test code to construct "DataSource" without creating
a Dataset object.
